♻️ HXSGQX task: decompose brief command
