+++ .agentplane/tasks/202605282151-HXSGQX/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 HXSGQX task: Task brief command decomposition [202605282151-HXSGQX]
